+++ app.js
@@ -5,6 +5,9 @@ const app = express();
 const { getCategories } = require("./controllers/getCategories-controller");
 const { getReviewById } = require("./controllers/getReviewById-controller");
 const { getReviews } = require("./controllers/getReviews-controller");
+const {
+  getReviewComments,
+} = require("./controllers/getReviewComments-controller");
 
 app.get("/api", (req, res) => {
   res.status(200).send({ message: "all okay" });
@@ -15,6 +18,8 @@ app.get("/api/categories", getCategories);
 app.get("/api/reviews/:review_id", getReviewById);
 
 app.get("/api/reviews", getReviews);
+
+app.get("/api/reviews/:review_id/comments", getReviewComments);
 
 app.use((err, req, res, next) => {
   if (err.status === 404) {
