+++ example.txt
@@ -1,18 +1,8 @@
 You are now connected to database "nc_games_test" as user "karllyuter".
-                        title                         |        designer        |      owner      |                                 review_img_url                                 |     category     |       created_at        | votes | review_id | comment_count 
-------------------------------------------------------+------------------------+-----------------+--------------------------------------------------------------------------------+------------------+-------------------------+-------+-----------+---------------
- Mollit elit qui incididunt veniam occaecat cupidatat | Avery Wunzboogerz      | mallionaire     | https://images.pexels.com/photos/776657/pexels-photo-776657.jpeg?w=700&h=700   | social deduction | 2021-01-25 11:16:54.963 |     9 |         7 |             0
- Dolor reprehenderit                                  | Gamey McGameface       | mallionaire     | https://images.pexels.com/photos/278918/pexels-photo-278918.jpeg?w=700&h=700   | social deduction | 2021-01-22 11:35:50.936 |     7 |         4 |             0
- Scythe; you're gonna need a bigger table!            | Jamey Stegmaier        | mallionaire     | https://images.pexels.com/photos/4200740/pexels-photo-4200740.jpeg?w=700&h=700 | social deduction | 2021-01-22 10:37:04.839 |   100 |        12 |             0
- Build you own tour de Yorkshire                      | Asger Harding Granerud | mallionaire     | https://images.pexels.com/photos/258045/pexels-photo-258045.jpeg?w=700&h=700   | social deduction | 2021-01-18 10:01:41.251 |    10 |        10 |             0
- Ultimate Werewolf                                    | Akihisa Okui           | bainesface      | https://images.pexels.com/photos/5350049/pexels-photo-5350049.jpeg?w=700&h=700 | social deduction | 2021-01-18 10:01:41.251 |     5 |         3 |             3
- Jenga                                                | Leslie Scott           | philippaclaire9 | https://images.pexels.com/photos/4473494/pexels-photo-4473494.jpeg?w=700&h=700 | dexterity        | 2021-01-18 10:01:41.251 |     5 |         2 |             3
- A truly Quacking Game; Quacks of Quedlinburg         | Wolfgang Warsch        | mallionaire     | https://images.pexels.com/photos/279321/pexels-photo-279321.jpeg?w=700&h=700   | social deduction | 2021-01-18 10:01:41.251 |    10 |         9 |             0
- That's just what an evil person would say!           | Fiona Lohoar           | mallionaire     | https://images.pexels.com/photos/220057/pexels-photo-220057.jpeg?w=700&h=700   | social deduction | 2021-01-18 10:01:41.251 |     8 |        11 |             0
- One Night Ultimate Werewolf                          | Akihisa Okui           | mallionaire     | https://images.pexels.com/photos/5350049/pexels-photo-5350049.jpeg?w=700&h=700 | social deduction | 2021-01-18 10:01:41.251 |     5 |         8 |             0
- Agricola                                             | Uwe Rosenberg          | mallionaire     | https://images.pexels.com/photos/974314/pexels-photo-974314.jpeg?w=700&h=700   | euro game        | 2021-01-18 10:00:20.514 |     1 |         1 |             0
- Proident tempor et.                                  | Seymour Buttz          | mallionaire     | https://images.pexels.com/photos/209728/pexels-photo-209728.jpeg?w=700&h=700   | social deduction | 2021-01-07 09:06:08.077 |     5 |         5 |             0
- Occaecat consequat officia in quis commodo.          | Ollie Tabooger         | mallionaire     | https://images.pexels.com/photos/207924/pexels-photo-207924.jpeg?w=700&h=700   | social deduction | 2020-09-13 15:19:28.077 |     8 |         6 |             0
- Settlers of Catan: Don't Settle For Less             | Klaus Teuber           | mallionaire     | https://images.pexels.com/photos/1153929/pexels-photo-1153929.jpeg?w=700&h=700 | social deduction | 1970-01-10 03:08:38.4   |    16 |        13 |             0
-(13 rows)
+ comment_id |                                                       body                                                       | review_id |     author      | votes |       created_at       
+------------+------------------------------------------------------------------------------------------------------------------+-----------+-----------------+-------+------------------------
+          6 | Not sure about dogs, but my cat likes to get involved with board games, the boxes are their particular favourite |         3 | philippaclaire9 |    10 | 2021-03-27 19:49:48.11
+          3 | I didn't know dogs could play games                                                                              |         3 | philippaclaire9 |    10 | 2021-01-18 10:09:48.11
+          2 | My dog loved this game too!                                                                                      |         3 | mallionaire     |    13 | 2021-01-18 10:09:05.41
+(3 rows)
